ssh_tunnel: lift relay helper into shared mngr_forward.relay module
Both the desktop client's reverse-tunnel handler and mngr_forward's
direct-tcpip forward path used near-identical _relay_step / _relay_data
functions. The recent EOF-detection fix had to be applied in two places;
duplicating that risk again would be a recurring source of drift bugs.

Lift the helper to libs/mngr_forward/imbue/mngr_forward/relay.py and
import it from both call sites. Adds a workspace dep edge from
apps/minds onto imbue-mngr-forward (only cost; relay is a paramiko
utility and mngr_forward is its long-term home per the spec).
