+++ apps/minds/imbue/minds/desktop_client/ssh_tunnel_test.py
@@ -13,9 +13,9 @@ from imbue.minds.desktop_client.ssh_tunnel import ReverseTunnelInfo
 from imbue.minds.desktop_client.ssh_tunnel import SSHTunnelError
 from imbue.minds.desktop_client.ssh_tunnel import SSHTunnelManager
 from imbue.minds.desktop_client.ssh_tunnel import _ForwardedTunnelHandler
-from imbue.minds.desktop_client.ssh_tunnel import _relay_data
 from imbue.minds.desktop_client.ssh_tunnel import _ssh_connection_is_active
 from imbue.minds.desktop_client.ssh_tunnel import _ssh_connection_transport
+from imbue.mngr_forward.relay import relay_data
 
 
 class FakeChannelFromSocket:
@@ -145,7 +145,7 @@ def test_ssh_connection_transport_raises_when_none() -> None:
         _ssh_connection_transport(client)
 
 
-# -- _relay_data tests --
+# -- relay_data tests --
 
 
 def test_relay_data_forwards_between_socket_pair() -> None:
@@ -154,7 +154,7 @@ def test_relay_data_forwards_between_socket_pair() -> None:
     channel_sock, relay_sock_b = socket.socketpair(socket.AF_UNIX, socket.SOCK_STREAM)
 
     fake_channel = FakeChannelFromSocket.create(relay_sock_b)
-    relay_thread = threading.Thread(target=_relay_data, args=(relay_sock_a, fake_channel), daemon=True)
+    relay_thread = threading.Thread(target=relay_data, args=(relay_sock_a, fake_channel), daemon=True)
     relay_thread.start()
 
     app_sock.settimeout(3.0)
@@ -168,74 +168,6 @@ def test_relay_data_forwards_between_socket_pair() -> None:
     app_sock.close()
     channel_sock.close()
     relay_thread.join(timeout=5.0)
-
-
-class FakeEofChannel:
-    """Stub paramiko-channel-like that simulates the half-closed (EOF-received) state.
-
-    paramiko marks ``Channel.fileno()`` as readable for any channel event
-    (data, EOF, window-adjust). The real paramiko channel after the remote
-    sends EOF reports ``recv_ready() is False`` and ``eof_received is True``,
-    yet ``select`` still wakes up on the fileno. This stub reproduces that
-    state by wrapping a Unix socket whose peer has been closed: ``select``
-    will mark it readable indefinitely. The relay must detect the EOF and
-    terminate, otherwise the loop spins at ~1M iters/sec.
-    """
-
-    _sock: socket.socket
-
-    @classmethod
-    def create(cls) -> "FakeEofChannel":
-        sock_a, sock_b = socket.socketpair(socket.AF_UNIX, socket.SOCK_STREAM)
-        sock_b.close()
-        instance = cls.__new__(cls)
-        object.__setattr__(instance, "_sock", sock_a)
-        return instance
-
-    def fileno(self) -> int:
-        return self._sock.fileno()
-
-    def recv_ready(self) -> bool:
-        return False
-
-    @property
-    def eof_received(self) -> bool:
-        return True
-
-    @property
-    def closed(self) -> bool:
-        return False
-
-    def recv(self, size: int) -> bytes:
-        return b""
-
-    def sendall(self, data: bytes) -> None:
-        pass
-
-    def close(self) -> None:
-        self._sock.close()
-
-
-def test_relay_data_terminates_when_channel_has_received_eof() -> None:
-    """Regression: half-closed channel (EOF received) must not spin the relay loop.
-
-    Before the fix, ``_relay_step`` fell through when ``select`` reported the
-    channel readable but ``recv_ready()`` was False, causing the relay thread
-    to spin at hundreds of thousands of iters/sec and pin a CPU core. The
-    fix detects ``eof_received``/``closed`` in that branch and terminates.
-    """
-    sock_a, sock_b = socket.socketpair(socket.AF_UNIX, socket.SOCK_STREAM)
-    eof_channel = FakeEofChannel.create()
-    relay_thread = threading.Thread(target=_relay_data, args=(sock_a, eof_channel), daemon=True)
-    relay_thread.start()
-    relay_thread.join(timeout=3.0)
-    try:
-        assert not relay_thread.is_alive(), (
-            "relay thread should have terminated on EOF-received channel; "
-            "without the fix it spins forever"
-        )
-    finally:
-        sock_b.close()
 
 
 # -- ReverseTunnelInfo --
