+++ apps/minds/pyproject.toml
@@ -15,6 +15,7 @@ dependencies = [
     "jinja2>=3.1.0",
     "loguru",
     "imbue-mngr",
+    "imbue-mngr-forward",
     "modal>=1.3.1",
     "paramiko>=3.0",
     "playwright>=1.40.0",
@@ -40,6 +41,7 @@ build-backend = "hatchling.build"
 [tool.uv.sources]
 concurrency-group = { workspace = true }
 imbue-mngr = { workspace = true }
+imbue-mngr-forward = { workspace = true }
 imbue-common = { workspace = true }
 
 [tool.hatch.build.targets.wheel]
